Refactor extraction functions into subdirectory
